Implemented sections in recipe

+++ app/controllers/recipes_controller.rb
@@ -18,10 +18,12 @@ class RecipesController < ApplicationController
   def new
     @recipe = Recipe.new
     @ingredients = []
+    @sections = []
   end
 
   def edit
     set_ingredients
+    set_sections
   end
 
   def create
@@ -34,14 +36,8 @@ class RecipesController < ApplicationController
         @recipe.errors.full_messages.each do |e|
           flash.now[:error] = e
         end
-        @ingredients = []
-        if recipe_params[:ingredients_recipes_attributes]
-          for i in recipe_params[:ingredients_recipes_attributes].values do
-            if i[:_destroy] != 1
-              @ingredients.append(i.except(:_destroy))
-            end
-          end
-        end
+        @ingredients = recipe_params.has_key?(:ingredients_recipes_attributes) ? recipe_params[:ingredients_recipes_attributes].values : []
+        @sections = recipe_params.has_key?(:sections_attributes) ? recipe_params[:sections_attributes].values : []
         format.html { render :new, status: :unprocessable_entity }
       end
     end
@@ -60,20 +56,16 @@ class RecipesController < ApplicationController
   def update 
     respond_to do |format|
       params = recipe_params
-      if !params.key?(:category_ids)
-        params[:category_ids] = []
-      end
-      if !params.key?(:ingredients_recipes_attributes)
-        IngredientsRecipe.where(recipe_id: @recipe.id).destroy_all
-        params[:ingredients_recipes_attributes] = {}
-      end
+      params = validate_params(params)
       @recipe.attributes = params
-      if @recipe.valid? && params[:ingredients_recipes_attributes] != {}
+      if @recipe.valid? && params[:ingredients_recipes_attributes] != {} && params[:sections_attributes] != {}
         IngredientsRecipe.where(recipe_id: @recipe.id).destroy_all
+        Section.where(recipe_id: @recipe.id).destroy_all
         @recipe.save
         format.html { redirect_to @recipe, notice: t(".notice") }
       else
         set_ingredients
+        set_sections
         @recipe.errors.full_messages.each do |e|
           flash.now[:error] = e
         end
@@ -98,8 +90,9 @@ class RecipesController < ApplicationController
   end
 
   def recipe_params
-    params.require(:recipe).permit(:title, :preparation_description, :time_in_minutes_needed, :difficulty, :user_id, :layout, category_ids: [], 
-      ingredients_recipes_attributes: [:id, :ingredient_name, :quantity, :unit, :_destroy])
+    params.require(:recipe).permit(:title, :time_in_minutes_needed, :difficulty, :user_id, :layout, category_ids: [], 
+      sections_attributes: [:id, :title, :body],
+      ingredients_recipes_attributes: [:id, :ingredient_name, :quantity, :unit])
   end
 
   def cook_books_params
@@ -108,9 +101,30 @@ class RecipesController < ApplicationController
 
   def set_ingredients
     @ingredients = []
-    for i in IngredientsRecipe.where("recipe_id=#{@recipe.id}") do
+    for i in IngredientsRecipe.where(recipe_id: @recipe.id) do
       j = { ingredient_name: i.ingredient.name, quantity: i.quantity, unit: IngredientsRecipe.units[i.unit] }
       @ingredients.append(j)
     end
+  end
+
+  def set_sections
+    @sections = []
+    for i in Section.where(recipe_id: @recipe.id) do
+      j = { title: i.title, body: i.body }
+      @sections.append(j)
+    end
+  end
+
+  def validate_params(params)
+    if !params.key?(:category_ids)
+      params[:category_ids] = []
+    end
+    if !params.key?(:ingredients_recipes_attributes)
+      params[:ingredients_recipes_attributes] = {}
+    end
+    if !params.key?(:sections_attributes)
+      params[:sections_attributes] = {}
+    end
+    params
   end
 end
